requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
google-cloud-secret-manager==2.21.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
httpx==0.28.0
pydantic-settings==2.7.0
